refactor: domain -> entity 이름 변경
