+++ src/main/java/com/thekey/stylekeyserver/brand/repository/BrandRepository.java
@@ -1,7 +1,7 @@
 package com.thekey.stylekeyserver.brand.repository;
 
-import com.thekey.stylekeyserver.brand.domain.Brand;
-import com.thekey.stylekeyserver.stylepoint.domain.StylePoint;
+import com.thekey.stylekeyserver.brand.entity.Brand;
+import com.thekey.stylekeyserver.stylepoint.entity.StylePoint;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
