+++ src/main/java/com/thekey/stylekeyserver/brand/service/BrandAdminService.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.brand.service;
 
-import com.thekey.stylekeyserver.brand.domain.Brand;
+import com.thekey.stylekeyserver.brand.entity.Brand;
 import com.thekey.stylekeyserver.brand.dto.request.BrandRequest;
 import com.thekey.stylekeyserver.brand.dto.response.BrandPageResponse;
 import java.util.List;
