+++ src/main/java/com/thekey/stylekeyserver/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.category.repository;
 
-import com.thekey.stylekeyserver.category.domain.Category;
+import com.thekey.stylekeyserver.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
