+++ src/main/java/com/thekey/stylekeyserver/category/service/CategoryService.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.category.service;
 
-import com.thekey.stylekeyserver.category.domain.Category;
+import com.thekey.stylekeyserver.category.entity.Category;
 import java.util.List;
 
 public interface CategoryService {
