+++ src/main/java/com/thekey/stylekeyserver/common/s3/service/S3Service.java
@@ -8,8 +8,8 @@ import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.thekey.stylekeyserver.common.exception.ApiException;
 import com.thekey.stylekeyserver.common.exception.ErrorCode;
-import com.thekey.stylekeyserver.image.domain.Image;
-import com.thekey.stylekeyserver.image.domain.Type;
+import com.thekey.stylekeyserver.image.entity.Image;
+import com.thekey.stylekeyserver.image.entity.Type;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
