+++ src/main/java/com/thekey/stylekeyserver/coordinatelook/dto/response/ApiCoordinateLookDetailsResponse.java
@@ -4,8 +4,8 @@ package com.thekey.stylekeyserver.coordinatelook.dto.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.thekey.stylekeyserver.coordinatelook.domain.CoordinateLook;
-import com.thekey.stylekeyserver.item.domain.Item;
+import com.thekey.stylekeyserver.coordinatelook.entity.CoordinateLook;
+import com.thekey.stylekeyserver.item.entity.Item;
 import com.thekey.stylekeyserver.item.dto.response.ItemResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
