+++ src/main/java/com/thekey/stylekeyserver/coordinatelook/dto/response/CoordinateLookResponse.java
@@ -2,7 +2,7 @@ package com.thekey.stylekeyserver.coordinatelook.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.thekey.stylekeyserver.coordinatelook.domain.CoordinateLook;
+import com.thekey.stylekeyserver.coordinatelook.entity.CoordinateLook;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Data;
