+++ src/main/java/com/thekey/stylekeyserver/image/service/ImageBatchService.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.image.service;
 
-import com.thekey.stylekeyserver.image.domain.Image;
+import com.thekey.stylekeyserver.image.entity.Image;
 import com.thekey.stylekeyserver.image.repository.ImageRepository;
 import com.thekey.stylekeyserver.common.s3.service.S3Service;
 import java.io.UnsupportedEncodingException;
