+++ src/main/java/com/thekey/stylekeyserver/image/service/ImageService.java
@@ -1,10 +1,8 @@
 package com.thekey.stylekeyserver.image.service;
 
 import com.thekey.stylekeyserver.common.s3.service.S3Service;
-import com.thekey.stylekeyserver.image.domain.Image;
+import com.thekey.stylekeyserver.image.entity.Image;
 import com.thekey.stylekeyserver.image.repository.ImageRepository;
-import java.io.UnsupportedEncodingException;
-import java.net.MalformedURLException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
