+++ src/main/java/com/thekey/stylekeyserver/item/service/ItemService.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.item.service;
 
-import com.thekey.stylekeyserver.item.domain.Item;
+import com.thekey.stylekeyserver.item.entity.Item;
 import com.thekey.stylekeyserver.item.dto.response.ApiItemResponse;
 import com.thekey.stylekeyserver.like.service.LikeItemService;
 import java.util.List;
