+++ src/main/java/com/thekey/stylekeyserver/stylepoint/repository/StylePointRepository.java
@@ -1,6 +1,6 @@
 package com.thekey.stylekeyserver.stylepoint.repository;
 
-import com.thekey.stylekeyserver.stylepoint.domain.StylePoint;
+import com.thekey.stylekeyserver.stylepoint.entity.StylePoint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
