+++ src/main/java/com/thekey/stylekeyserver/test/entity/TestResult.java
@@ -2,7 +2,7 @@ package com.thekey.stylekeyserver.test.entity;
 
 import com.thekey.stylekeyserver.auth.entity.User;
 import com.thekey.stylekeyserver.base.BaseTimeEntity;
-import com.thekey.stylekeyserver.stylepoint.domain.StylePoint;
+import com.thekey.stylekeyserver.stylepoint.entity.StylePoint;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,7 +13,6 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import org.hibernate.annotations.BatchSize;
 
 @Entity
 @Table(name = "test_result")
